+++ saleor/graphql/discount/mutations/sale/sale_remove_catalogues.py
@@ -5,7 +5,6 @@ from .....graphql.channel import ChannelContext
 from .....permission.enums import DiscountPermissions
 from .....webhook.event_types import WebhookEventAsyncType
 from ....core import ResolveInfo
-from ....core.descriptions import DEPRECATED_IN_3X_MUTATION
 from ....core.doc_category import DOC_CATEGORY_DISCOUNTS
 from ....core.types import DiscountError
 from ....core.utils import (
@@ -24,11 +23,7 @@ from .sale_base_catalogue import SaleBaseCatalogueMutation
 
 class SaleRemoveCatalogues(SaleBaseCatalogueMutation):
     class Meta:
-        description = (
-            "Removes products, categories, collections from a sale."
-            + DEPRECATED_IN_3X_MUTATION
-            + " Use `promotionRuleUpdate` or `promotionRuleDelete` mutations instead."
-        )
+        description = "Removes products, categories, collections from a sale."
         doc_category = DOC_CATEGORY_DISCOUNTS
         permissions = (DiscountPermissions.MANAGE_DISCOUNTS,)
         error_type_class = DiscountError