+++ saleor/graphql/meta/mutations/update_private_metadata.py
@@ -32,11 +32,17 @@ class UpdatePrivateMetadata(BaseMetadataMutation):
     @classmethod
     def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
         instance = cls.get_instance(info, **data)
+
         if instance:
             meta_instance = get_valid_metadata_instance(instance)
             metadata_list = data.pop("input")
-            cls.validate_metadata_keys(metadata_list)
+
+            cls.create_metadata_from_graphql_input(
+                metadata_list, error_field_name="input"
+            )
+
             items = {data.key: data.value for data in metadata_list}
             meta_instance.store_value_in_private_metadata(items=items)
             update_private_metadata(meta_instance, items)
+
         return cls.success_response(instance)