+++ saleor/graphql/order/schema.py
@@ -12,7 +12,7 @@ from ..core.connection import (
     filter_connection_queryset,
 )
 from ..core.context import get_database_connection_name
-from ..core.descriptions import DEPRECATED_IN_3X_FIELD
+from ..core.descriptions import DEFAULT_DEPRECATION_REASON
 from ..core.doc_category import DOC_CATEGORY_ORDERS
 from ..core.enums import ReportingPeriod
 from ..core.fields import (
@@ -103,7 +103,7 @@ class OrderQueries(graphene.ObjectType):
         permissions=[
             OrderPermissions.MANAGE_ORDERS,
         ],
-        deprecation_reason=DEPRECATED_IN_3X_FIELD,
+        deprecation_reason=DEFAULT_DEPRECATION_REASON,
     )
     order = BaseField(
         Order,
@@ -161,12 +161,12 @@ class OrderQueries(graphene.ObjectType):
             OrderPermissions.MANAGE_ORDERS,
         ],
         doc_category=DOC_CATEGORY_ORDERS,
-        deprecation_reason=DEPRECATED_IN_3X_FIELD,
+        deprecation_reason=DEFAULT_DEPRECATION_REASON,
     )
     order_by_token = BaseField(
         Order,
         description="Look up an order by token.",
-        deprecation_reason=DEPRECATED_IN_3X_FIELD,
+        deprecation_reason=DEFAULT_DEPRECATION_REASON,
         token=graphene.Argument(UUID, description="The order's token.", required=True),
         doc_category=DOC_CATEGORY_ORDERS,
     )
@@ -259,12 +259,12 @@ class OrderMutations(graphene.ObjectType):
     draft_order_delete = DraftOrderDelete.Field()
     draft_order_bulk_delete = DraftOrderBulkDelete.Field()
     draft_order_lines_bulk_delete = DraftOrderLinesBulkDelete.Field(
-        deprecation_reason=DEPRECATED_IN_3X_FIELD
+        deprecation_reason=DEFAULT_DEPRECATION_REASON
     )
     draft_order_update = DraftOrderUpdate.Field()
 
     order_add_note = OrderAddNote.Field(
-        deprecation_reason=(f"{DEPRECATED_IN_3X_FIELD} Use `orderNoteAdd` instead.")
+        deprecation_reason="Use `orderNoteAdd` instead."
     )
     order_cancel = OrderCancel.Field()
     order_capture = OrderCapture.Field()