+++ saleor/graphql/payment/mutations/transaction/transaction_create.py
@@ -36,7 +36,7 @@ from ....meta.inputs import MetadataInput
 from ....plugins.dataloaders import get_plugin_manager_promise
 from ...enums import TransactionActionEnum
 from ...types import TransactionItem
-from ...utils import metadata_contains_empty_key
+from ...utils import deprecated_metadata_contains_empty_key
 from ..payment.payment_check_balance import MoneyInput
 
 
@@ -125,13 +125,15 @@ class TransactionCreate(BaseMutation):
                 }
             ) from e
 
+    # TODO This should be unified with metadata_manager and MetadataItemCollection
+    # EXT-2054
     @classmethod
-    def validate_metadata_keys(  # type: ignore[override]
+    def validate_metadata_keys(
         cls, metadata_list: list[dict] | None, field_name, error_code
     ):
         if not metadata_list:
             return
-        if metadata_contains_empty_key(metadata_list):
+        if deprecated_metadata_contains_empty_key(metadata_list):
             raise ValidationError(
                 {
                     "transaction": ValidationError(